backend/notifications/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aioodbc==0.5.0
python-decouple==3.8
pydantic-settings==2.0.3
